requirements.txt
Flask
flask-login
flask-cors
mysql-connector
names
